requirements.txt
yt-dlp
fastapi>=0.111.0
uvicorn>=0.30.0
websockets>=12.0
pydantic>=2.7.0
psutil>=5.8.0
pywebview>=5.0.0
pyinstaller>=6.5.0
